created a basic frontend display logic for post and comments need to improve them and do not touch api calls they are correct and will be improved later when neeeded

+++ Services/Post/src/graphql/resolvers/mutations/Dislike/index.ts
@@ -1,13 +1,25 @@
-import { Comment, Dislike, Post } from "../../../../models";
+import { Comment, Dislike, Like, Post } from "../../../../models";
+import { removeLike } from "../Like";
 
 const addDislike = async (on: string, by: string) => {
   try {
+    const dislikeUpdata =
+      (await Post.findById(on)) || (await Comment.findById(on));
+    if (!dislikeUpdata) throw new Error("Provide a valid ID");
+    const liked = await Like.findOne({ on: on, likedBy: by });
+    if (liked) {
+      await removeLike(on, by);
+    }
     const check = await Dislike.findOne({ on: on, dislikedBy: by });
-    if (check) throw new Error("Error performing dislike operation");
+    if (check) {
+      return await removeDislike(on, by);
+    }
     const dislike = await Dislike.create({
       on: on,
       dislikedBy: by,
     });
+    dislikeUpdata.dislikeCount += 1;
+    await dislikeUpdata.save();
     return dislike.save();
   } catch (err) {
     throw new Error("Failed to dislike");
@@ -16,6 +28,9 @@ const addDislike = async (on: string, by: string) => {
 
 const removeDislike = async (on: string, by: string) => {
   try {
+    const dislikeUpdata =
+      (await Post.findById(on)) || (await Comment.findById(on));
+    if (!dislikeUpdata) throw new Error("Provide a valid ID");
     const dislike = await Dislike.findOneAndDelete({
       on: on,
       dislikedBy: by,
@@ -23,6 +38,8 @@ const removeDislike = async (on: string, by: string) => {
     if (!dislike) {
       throw new Error("Dislike not found");
     }
+    dislikeUpdata.dislikeCount -= 1;
+    await dislikeUpdata.save();
     return "Dislike removed successfully";
   } catch (err) {
     throw new Error("Failed to remove dislike");
